requirements.txt
fastapi
uvicorn
pymongo
motor
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
pydantic[email]